[~] Переделывание файловой структуры
R: Бизнес логика была в приложении

FB: Вынесена в корневую папку
