+++ orders/templates/orders/includes/navbar.html
@@ -22,10 +22,10 @@
 
         </div>
         <div class="col-6">
-            <nav class="ms-5">
-                <a class="p-2 btn btn-outline-dark" href="{% url 'orders:orders_list_html' %}">Список заказов</a>
-                <a class="p-2 btn btn-outline-dark" href="{% url 'dishes:dish_list_html' %}">Список блюд</a>
-                <a class="p-2 btn btn-outline-dark" href="{% url 'orders:table_list_html' %}">Список столов</a>
+            <nav class="ms-5" style="text-align: right">
+                <a class="p-2 btn btn-outline-dark" href="{% url 'orders:orders_list_html' %}">Заказы</a>
+                <a class="p-2 btn btn-outline-dark" href="{% url 'dishes:dish_list_html' %}">Блюда</a>
+                <a class="p-2 btn btn-outline-dark" href="{% url 'orders:table_list_html' %}">Столы</a>
                 <a class="p-2 btn btn-outline-warning" href="{% url 'orders:orders_revenue_html' %}">Расчет выручки</a>
                 <form action="{% url 'orders:logout' %}" method="post" style="display: inline-block;">
                     {% csrf_token %}
