+++ orders/templates/orders/order_detail.html
@@ -1,7 +1,7 @@
 {% extends 'orders/base.html' %}
 {% block content %}
 <div class="pricing-header px-3 py-3 pt-md-5 pb-md-4 mx-auto text-center">
-    <h1 class="display-4">{{order}}</h1>
+    <h1 class="display-6">{{order}}</h1>
 </div>
 
 <div class="container">
@@ -15,7 +15,7 @@
                 <div class="card-body">
                     <p>
                     <h5>Дата заказа:</h5>
-                    {{ order.datetime }}
+                    {{ order.order_datetime }}
                     <p>
                     <h5>Номер стола:</h5>
                     {{ order.table_number }}
