+++ orders/templates/orders/search_results.html
@@ -6,6 +6,12 @@
     <div class="row text-center">
 
         <div class="row">
+            <div class="row">
+            <div class="pricing-header px-3 py-3 pt-md-5 pb-md-4 mx-auto text-center">
+            <h3 class="display-6">
+                Результаты поиска
+            </h3>
+        </div>
             <p>
                 <a class="btn btn-lg btn-block btn-outline-primary"
                    href="{% url 'orders:orders_list_html' %}"
