+++ orders/templates/orders/table_form.html
@@ -1,7 +1,7 @@
 {% extends 'orders/base.html' %}
 {% block content %}
 <div class="pricing-header px-3 py-3 pt-md-5 pb-md-4 mx-auto text-center">
-    <h1 class="display-4">
+    <h1 class="display-6">
         {% if object %}
         Изменение информации о столе
         {% else %}
