+++ orders/templates/orders/table_list.html
@@ -4,6 +4,11 @@
 
 <div class="container">
     <div class="row text-center">
+        <div class="pricing-header px-3 py-3 pt-md-5 pb-md-4 mx-auto text-center">
+            <h3 class="display-6">
+                Список всех столов
+            </h3>
+        </div>
         <div class="row">
             <p>
                 <a class="btn btn-lg btn-block btn-outline-success"
