+++ orders/view_html.py
@@ -7,9 +7,10 @@ from django.shortcuts import render
 from django.urls import reverse_lazy
 from django.views.generic import CreateView, DeleteView, DetailView, ListView, UpdateView
 
-from order_logic.services import OrderService, PaginationService, number_of_lines
+from logics.order_logics import OrderService
 from orders.forms import OrderCreateForm, OrderUpdateForm, TableForm
 from orders.models import Order, Table
+from services.display_html import PaginationService, number_of_lines
 
 
 class OrderListView(LoginRequiredMixin, ListView):
