+++ orders/views_api.py
@@ -7,7 +7,7 @@ from rest_framework.response import Response
 from rest_framework.serializers import Serializer
 from rest_framework.views import APIView
 
-from order_logic.services import OrderService
+from logics.order_logics import OrderService
 from orders.filters import OrderFilter
 from orders.models import Order, Table
 from orders.serializers import (
